core: Move Freezer class to tools namespace
